refactor: reworked page nav
