+++ src/app/app/page.tsx
@@ -2,13 +2,13 @@
 
 import Link from "next/link";
 import { PageNav } from "@/components/PageNav/PageNav";
-import { redirect } from "next/navigation";
+import { AppPage } from "@/shared/entities";
 
 const App = () => {
     return (
     <div className="flex min-h-screen flex-col items-center justify-center p-32">
-        <PageNav charId={10}/>
-        <Link className="text-center bg-purple-50 text-gray-700 rounded p-2" href={"/app/characters/10"}>Home</Link>
+        <PageNav charId={0} highlighted={AppPage.DETAILS}/>
+        <Link className="text-center bg-purple-50 text-gray-700 rounded p-2" href={"/app/characters/0"}>Home</Link>
     </div>
     );
 }
